Abstract out functions from renderer.ts to cleanup.
